Creacion de nueva tarea con la segunda ventana

+++ app/src/main/java/com/aplicacionTask/data/local/entity/Task.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "task")
 data class Task(
-    @PrimaryKey(autoGenerate = true) val id : Int? = null,
+    @PrimaryKey(autoGenerate = true) val id : Long? = null,
     @ColumnInfo(name = "title") val title : String? = null,
     @ColumnInfo(name = "description") val description : String? = null
 )
